refactor: 예외 핸들러를 하나로 통합
