+++ src/main/java/roomescape/controller/exception/GlobalExceptionHandler.java
@@ -2,8 +2,7 @@ package roomescape.controller.exception;
 
 import com.fasterxml.jackson.databind.exc.InvalidFormatException;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.core.Ordered;
-import org.springframework.core.annotation.Order;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.web.bind.MissingPathVariableException;
@@ -11,12 +10,79 @@ import org.springframework.web.bind.MissingServletRequestParameterException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
+import roomescape.exception.CodeException;
+import roomescape.exception.DuplicateReservationException;
+import roomescape.exception.EntityNotFoundException;
 import roomescape.exception.ErrorCode;
+import roomescape.exception.InUseEntityException;
+import roomescape.exception.InvalidInputException;
+import roomescape.exception.NotAcceptableReservationException;
+
 
 @RestControllerAdvice
 @Slf4j
-@Order(Ordered.HIGHEST_PRECEDENCE)
-public class HttpRequestExceptionHandler {
+public class GlobalExceptionHandler {
+
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<ErrorResponse> handleInternalServerError(Exception exception) {
+        log.error("[Internal Server Error]", exception);
+
+        ErrorResponse response = new ErrorResponse("예상하지 못한 예외가 발생했습니다.", ErrorCode.INTERNAL_SERVER_ERROR);
+
+        return ResponseEntity.internalServerError()
+                .body(response);
+    }
+
+    @ExceptionHandler({
+            IllegalArgumentException.class,
+            IllegalStateException.class
+    })
+    public ResponseEntity<ErrorResponse> handleDefaultBadRequest(Exception exception) {
+        log.warn("[Bad Request]", exception);
+        ErrorResponse response = new ErrorResponse(exception.getMessage(), ErrorCode.BAD_REQUEST);
+
+        return ResponseEntity.badRequest()
+                .body(response);
+    }
+
+    @ExceptionHandler(InvalidInputException.class)
+    public ResponseEntity<ErrorResponse> handleBadRequest(InvalidInputException exception) {
+        log.warn("[Bad Request]", exception);
+        ErrorResponse response = new ErrorResponse(exception.getMessage(), exception.getErrorCode());
+
+        return ResponseEntity.badRequest()
+                .body(response);
+    }
+
+    @ExceptionHandler(NotAcceptableReservationException.class)
+    public ResponseEntity<ErrorResponse> handleForbidden(NotAcceptableReservationException exception) {
+        log.warn("[Forbidden]", exception);
+        ErrorResponse response = new ErrorResponse(exception.getMessage(), exception.getErrorCode());
+
+        return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                .body(response);
+    }
+
+    @ExceptionHandler(EntityNotFoundException.class)
+    public ResponseEntity<ErrorResponse> handleNotFound(EntityNotFoundException exception) {
+        log.warn("[Not Found]", exception);
+        ErrorResponse response = new ErrorResponse(exception.getMessage(), exception.getErrorCode());
+
+        return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                .body(response);
+    }
+
+    @ExceptionHandler({
+            InUseEntityException.class,
+            DuplicateReservationException.class
+    })
+    public ResponseEntity<ErrorResponse> handleConflict(CodeException exception) {
+        log.warn("[Conflict]", exception);
+        ErrorResponse response = new ErrorResponse(exception.getMessage(), exception.getErrorCode());
+
+        return ResponseEntity.status(HttpStatus.CONFLICT)
+                .body(response);
+    }
 
     @ExceptionHandler(MissingPathVariableException.class)
     public ResponseEntity<ErrorResponse> handleMissingPathVariable(
